Add links to Install and Tuturials
Podman does not have a great install.md document for it alone,
so link the install command to the tutorial for now that explains
how to install podman.  We should break this out into its own page.

install.md on libpod currectly describes how to checkout and build
libpod, it does not mention installing podman from rpm or apt.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>