Add mailing list button and page
Create a button on the home page for the mailing list info and
a page detailing how to subscribe.

"mailto" links are notoriously twitchy so I was hesitant to have
only a link to one on the main page.  I wanted to create a page
with one on it which also had an explanation on how to sign up for the list
via the web in case the mailto didn't work.  I also wanted to add
a warning that the list is not a replacement for "New issue" button on
GitHub.

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>