Merge lipod install.md to this repo
Merge the installation.md from Libpod's GitHub repo to this repo.  A link
will be left on the repo side of things to point to this merged file.
We've had a number of edits made to both files, sometimes not at the same
time and we'd a pretty good divergence in only a few months.

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>