Change URLs to skip redirects. Prefer HTTPS over HTTP. Use curl instead of wget.
Change URLs to URLs where all redirects have been followed (301 Moved Permanently).
For instance the URL:

https://download.opensuse.org/repositories/devel:kubic:libcontainers:testing/CentOS_7/devel:kubic:libcontainers:testing.repo

will after redirects be:

https://download.opensuse.org/repositories/devel:/kubic:/libcontainers:/testing/CentOS_7/devel:kubic:libcontainers:testing.repo

Use curl instead of wget.
At least in Centos 8, wget is not included in the "minimal install" set of packages.

Move the -o curl option to the beginning of the line.

Use the -L flag when using curl.

Signed-off-by: Erik Sjölund <erik.sjolund@gmail.com>