Replace podman-machine post
Anders created a post in early 2019 concerning his podman-machine project which has
since been deprecated in favor of `podman machine`.  As that post is at the top of the list when
searching in Google for `podman machine`, I've kept the page but replaced it's contents with
pointers for Podman Machine.

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>