fix broken windows download link (#523)
We should not use the latest link for github releases. It will will
always use the latest tag based on the date not semver. So this will
point to v3.4.7 at the moment.

Because such things can happen again we should manually set the version
and update it for each release. Ideally we have some form of automation
for this.

see https://github.com/containers/podman/issues/13947

Signed-off-by: Paul Holzinger <pholzing@redhat.com>